Type activité finalisé

+++ src/main/java/com/example/erpbackend/ServiceImplementation/TypeActiviteServiceImplement.java
@@ -16,11 +16,35 @@ import java.util.List;
 public class TypeActiviteServiceImplement implements TypeActiviteService {
 
     @Autowired
-    public  TypeActiviteRepository typeActiviteRepository;
+    public TypeActiviteRepository typeActiviteRepository;
+
     @Override
-    public Type_activite ajouterTypeActivite(Type_activite type_activite) {
+    public ReponseMessage ajouterTypeActivite(Type_activite type_activite) {
+
+        if (typeActiviteRepository.findByTypeActivite(type_activite.getTypeActivite()) == null)
+        {
+            typeActiviteRepository.save(type_activite);
+            ReponseMessage reponse = new ReponseMessage("Type activité ajouter avec success",true);
+            return reponse;
+        }
+        else{
+            ReponseMessage reponse = new ReponseMessage("Type activité existants",false);
+            return reponse;
+        }
+
+
+
+    }
+
+
+
+
+
+    /*public Type_activite ajouterTypeActivite(Type_activite type_activite) {
         return typeActiviteRepository.save(type_activite);
     }
+
+     */
 
     @Override
     public List<Type_activite> afficherTypeActivite() {
@@ -28,25 +52,42 @@ public class TypeActiviteServiceImplement implements TypeActiviteService {
     }
 
     @Override
-    public Type_activite modifierTypeActivite(Long id, Type_activite type_activite) {
+    public ReponseMessage modifierTypeActivite(Long id, Type_activite type_activite) {
 
+        if (typeActiviteRepository.findByIdactivite(id) != null){
 
+        return  typeActiviteRepository.findById(id)
+                    .map(type_activite1 -> {
+                        type_activite1.setTypeActivite(type_activite.getTypeActivite());
+                        type_activite1.setTypePostulant(type_activite.getTypePostulant());
+                         typeActiviteRepository.save(type_activite1);
+                        ReponseMessage rep = new ReponseMessage("Type activité modifier avec success",true);
+                          return rep;
+                    }).orElseThrow(() -> new RuntimeException("Type activité non trouvé 1"));
+        }
+        else {
+            ReponseMessage reponse = new ReponseMessage("Type activité non trouvé", false);
+            return reponse;
+        }
 
-        return typeActiviteRepository.findById(id)
-                .map(type_activite1 -> {
-                    type_activite1.setTypeActivite(type_activite.getTypeActivite());
-                    type_activite1.setTypePostulant(type_activite.getTypePostulant());
-                    return  typeActiviteRepository.save(type_activite1);
-                }).orElseThrow(() -> new RuntimeException("Type activité non trouvé"));
     }
 
-  /*  @Override
-    public String supprimerTypeActvite(Long id) {
-        return typeActiviteRepository.deleteById(id);
+    @Override
+    public ReponseMessage supprimerTypeActvite(Long id) {
+        if (typeActiviteRepository.findByIdactivite(id) != null) {
+            ReponseMessage rep;
+            typeActiviteRepository.deleteById(id);
+
+            rep = new ReponseMessage("Suppression effectuée", true);
+
+            return rep;
+        } else {
+            ReponseMessage rep = new ReponseMessage("Impossible de supprimer", false);
+
+            return rep;
+        }
+
+
     }
-
-   */
-
-
 }
